Frontend routes, pages, theme, general layout, user login
Implemented the foundation that the frontend will be built upon to improve future development speed.

As per my initial thoughts, all routes and pages have been created with a simple template.

Also implemented a simple user login to open up doors in the future for an admin dashboard / website editor.